Dodanie walidacji w delete_client_in_database

+++ src/client.py
@@ -45,4 +45,6 @@ class Client:
         return self.database.edit_client(self)
 
     def delete_client_in_database(self, id):
+        if not isinstance(id, int) or id < 0 or str(id) == 'True' or str(id) == 'False':
+            raise ValueError('Id musi być dodatnią liczbą całkowitą')
         return self.database.delete_client(id)
